Props and Prop Drilling 1. Parent to child data passing via Properties (Prop)
App.js -> Content.js -> ItemList.js -> Footer.js

2. Parent to children (Among siblings) data passing| Drilling via properties
