+++ react-app/src/Footer.js
@@ -1,12 +1,13 @@
 import React from 'react'
 
-const Footer = () => {
+const Footer = ({length}) => {
     const today = new Date();
   return (
     <footer>
-        <p>Copyright &copy; {today.getFullYear()}</p>
+        <p>Copyright &copy; {today.getFullYear()} | { length } List {length === 1 ? "Item" : "Items"}</p>
     </footer>
   )
 }
+
 
 export default Footer
