+++ react-app/src/LineItem.js
@@ -1,0 +1,28 @@
+/* this file is for list items */
+import { FaTrashAlt } from "react-icons/fa";
+
+const LineItem = ({item, handleCheck, handleDelete}) => {
+  return (
+    <li className="item" key={item.id}>
+        <input 
+            type="checkbox"
+            onChange={()=>handleCheck(item.id)} //anonymous func
+            checked={item.checked}
+        />
+        <label 
+            style={(item.checked) ? {textDecoration: "line-through"} : null}
+            onDoubleClick={()=>handleCheck(item.id)}
+        >{item.item}
+        </label>
+        {/*<button>Delete</button>*/}
+        <FaTrashAlt 
+            onClick={()=>handleDelete(item.id)}
+            role="button" 
+            tabIndex="0" 
+            aria-label={`Delete ${item.item}`}
+        />
+    </li>
+  )
+}
+
+export default LineItem
